refactor(filter): create filter package for security and fix @NonNull import
