+++ src/main/java/com/vire/virebackend/security/filter/JwtAuthenticationFilter.java
@@ -1,13 +1,15 @@
-package com.vire.virebackend.security;
+package com.vire.virebackend.security.filter;
 
 import com.vire.virebackend.repository.UserRepository;
+import com.vire.virebackend.security.CustomUserDetails;
+import com.vire.virebackend.security.JwtService;
 import io.jsonwebtoken.JwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
+import org.springframework.lang.NonNull;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.context.SecurityContextHolder;
