+++ src/main/java/com/vire/virebackend/security/filter/RequestSizeLimitFilter.java
@@ -1,10 +1,10 @@
-package com.vire.virebackend.security;
+package com.vire.virebackend.security.filter;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.jspecify.annotations.NonNull;
+import org.springframework.lang.NonNull;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
